+++ packages/bridge/lib/pi/plugin/event.ex
@@ -3,6 +3,8 @@ defmodule Pi.Plugin.Event do
 
   use GenServer
 
+  alias Pi.Transport.Stdio
+
   def start_link(opts \\ []), do: GenServer.start_link(__MODULE__, opts, name: __MODULE__)
 
   def install do
@@ -15,6 +17,10 @@ defmodule Pi.Plugin.Event do
   def push(event) when is_map(event) do
     install()
     GenServer.cast(__MODULE__, {:event, event})
+  end
+
+  def emit(name, data \\ %{}) when is_binary(name) and is_map(data) do
+    Stdio.emit(%{type: :event, name: name, data: data})
   end
 
   def recent(n \\ 50), do: GenServer.call(__MODULE__, {:recent, n})