+++ packages/bridge/test/pi/plugin/api_test.exs
@@ -4,10 +4,10 @@ defmodule Pi.Plugin.APITest do
   alias Pi.Plugin.API
 
   test "normalizes keyword and map API metadata" do
-    assert %API{name: :demo, module: __MODULE__, alias: :Demo} =
-             API.new(name: :demo, module: __MODULE__, alias: :Demo)
+    assert %API{name: :demo, module: __MODULE__, alias: :APITest} =
+             API.new(name: :demo, module: __MODULE__)
 
-    assert %API{name: :map_demo, module: __MODULE__} =
-             API.new(%{name: :map_demo, module: __MODULE__})
+    assert %API{name: :map_demo, module: __MODULE__, alias: :CustomAlias} =
+             API.new(%{name: :map_demo, module: __MODULE__, alias: :CustomAlias})
   end
 end