+++ packages/bridge/test/pi/plugin/manager_test.exs
@@ -11,9 +11,32 @@ defmodule Pi.Plugin.ManagerTest do
     end
   end
 
+  defmodule MacroDemo do
+    use Pi.Plugin
+
+    api(name: :macro_manager_demo, module: __MODULE__)
+    command(name: :macro_command, description: "Run macro command")
+
+    def handle_command(:macro_command, args, state), do: {{:ok, "ran #{args}"}, state}
+  end
+
+  defmodule HookDemo do
+    use Pi.Plugin
+
+    def tool_call(%{"input" => input}, _context, state) do
+      {{:ok, %{"input" => Map.put(input, "patched", true)}}, state}
+    end
+
+    def tool_result(_result, _context, state),
+      do: {{:ok, %{"content" => "patched result"}}, state}
+  end
+
   setup do
-    if pid = Process.whereis(Manager), do: GenServer.stop(pid)
-    if pid = Process.whereis(Pi.Plugin.Supervisor), do: DynamicSupervisor.stop(pid)
+    if pid = Process.whereis(Manager), do: stop_if_alive(pid, &GenServer.stop/1)
+
+    if pid = Process.whereis(Pi.Plugin.Supervisor),
+      do: stop_if_alive(pid, &DynamicSupervisor.stop/1)
+
     :ok
   end
 
@@ -24,6 +47,32 @@ defmodule Pi.Plugin.ManagerTest do
 
     assert [%Pi.Plugin.API{name: :manager_demo, module: Demo, alias: :ManagerDemo}] =
              Manager.apis()
+  end
+
+  test "loads and unloads plugins dynamically" do
+    {:ok, _pid} = Manager.start_link(plugins: [])
+
+    assert :ok = Manager.load(MacroDemo)
+    assert {:error, :already_loaded} = Manager.load(MacroDemo)
+    assert [%Pi.Protocol.PluginInfo{module: MacroDemo, name: "MacroDemo"}] = Manager.plugins()
+
+    assert [%Pi.Plugin.API{name: :macro_manager_demo, module: MacroDemo, alias: :MacroDemo}] =
+             Manager.apis()
+
+    assert [%Pi.Plugin.Command{name: :macro_command, plugin: MacroDemo}] = Manager.commands()
+    assert {:ok, "ran args"} = Manager.run_command(:macro_command, "args")
+
+    assert :ok = Manager.unload(MacroDemo)
+    assert [] = Manager.plugins()
+  end
+
+  test "runs plugin tool hook pipelines" do
+    {:ok, _pid} = Manager.start_link(plugins: [HookDemo])
+
+    assert {:ok, %{"input" => %{"patched" => true}}} =
+             Manager.tool_call(%{"input" => %{}}, %{})
+
+    assert {:ok, %{"content" => "patched result"}} = Manager.tool_result(%{}, %{})
   end
 
   test "restarts a plugin worker after process exit" do
@@ -50,4 +99,10 @@ defmodule Pi.Plugin.ManagerTest do
   end
 
   defp eventually(_fun, 0), do: false
+
+  defp stop_if_alive(pid, stop) do
+    if Process.alive?(pid), do: stop.(pid)
+  catch
+    :exit, _reason -> :ok
+  end
 end